@@ -24,24 +24,24 @@ public extension Entita2Entity {
         Self.IDAsKey(ID: getID())
     }
 
-    static func loadBy(IDBytes: Bytes, within transaction: AnyTransaction? = nil) async throws -> Self? {
+    static func loadBy(IDBytes: Bytes, within transaction: (any Transaction)? = nil) async throws -> Self? {
         try await Self.loadByRaw(
             IDBytes: Self.IDBytesAsKey(bytes: IDBytes)
         )
     }
 
-    static func loadByRaw(IDBytes: Bytes, within transaction: AnyTransaction? = nil) async throws -> Self? {
+    static func loadByRaw(IDBytes: Bytes, within transaction: (any Transaction)? = nil) async throws -> Self? {
         return try await self.afterLoadRoutines0(
             maybeBytes: try await self.storage.load(by: IDBytes, within: transaction),
             within: transaction
         )
     }
 
-    static func load(by ID: Identifier, within transaction: AnyTransaction? = nil) async throws -> Self? {
+    static func load(by ID: Identifier, within transaction: (any Transaction)? = nil) async throws -> Self? {
         try await Self.loadByRaw(IDBytes: Self.IDAsKey(ID: ID), within: transaction)
     }
 
-    static func afterLoadRoutines0(maybeBytes: Bytes?, within transaction: AnyTransaction?) async throws -> Self? {
+    static func afterLoadRoutines0(maybeBytes: Bytes?, within transaction: (any Transaction)?) async throws -> Self? {
         guard let bytes = maybeBytes else {
             return nil
         }
@@ -54,20 +54,20 @@ public extension Entita2Entity {
         return entity
     }
 
-    func afterLoad0(within transaction: AnyTransaction?) async throws {}
-    func afterLoad(within transaction: AnyTransaction?) async throws {}
-    func beforeSave0(within transaction: AnyTransaction?) async throws {}
-    func beforeSave(within transaction: AnyTransaction?) async throws {}
-    func afterSave0(within transaction: AnyTransaction?) async throws {}
-    func afterSave(within transaction: AnyTransaction?) async throws {}
-    func beforeInsert0(within transaction: AnyTransaction?) async throws {}
-    func beforeInsert(within transaction: AnyTransaction?) async throws {}
-    func afterInsert(within transaction: AnyTransaction?) async throws {}
-    func afterInsert0(within transaction: AnyTransaction?) async throws {}
-    func beforeDelete0(within transaction: AnyTransaction?) async throws {}
-    func beforeDelete(within transaction: AnyTransaction?) async throws {}
-    func afterDelete0(within transaction: AnyTransaction?) async throws {}
-    func afterDelete(within transaction: AnyTransaction?) async throws {}
+    func afterLoad0(within transaction: (any Transaction)?) async throws {}
+    func afterLoad(within transaction: (any Transaction)?) async throws {}
+    func beforeSave0(within transaction: (any Transaction)?) async throws {}
+    func beforeSave(within transaction: (any Transaction)?) async throws {}
+    func afterSave0(within transaction: (any Transaction)?) async throws {}
+    func afterSave(within transaction: (any Transaction)?) async throws {}
+    func beforeInsert0(within transaction: (any Transaction)?) async throws {}
+    func beforeInsert(within transaction: (any Transaction)?) async throws {}
+    func afterInsert(within transaction: (any Transaction)?) async throws {}
+    func afterInsert0(within transaction: (any Transaction)?) async throws {}
+    func beforeDelete0(within transaction: (any Transaction)?) async throws {}
+    func beforeDelete(within transaction: (any Transaction)?) async throws {}
+    func afterDelete0(within transaction: (any Transaction)?) async throws {}
+    func afterDelete(within transaction: (any Transaction)?) async throws {}
 
     func getPackedSelf() throws -> Bytes {
         let result: Bytes
@@ -84,7 +84,7 @@ public extension Entita2Entity {
     //MARK: - Public 0-methods
     
     /// This method is not intended to be used directly. Use `save()` method.
-    func save0(by ID: Identifier? = nil, within transaction: AnyTransaction?) async throws {
+    func save0(by ID: Identifier? = nil, within transaction: (any Transaction)?) async throws {
         let IDBytes: Bytes
         if let ID = ID {
             IDBytes = Self.IDAsKey(ID: ID)
@@ -100,7 +100,7 @@ public extension Entita2Entity {
     }
 
     /// This method is not intended to be used directly. Use `save()` method.
-    func delete0(within transaction: AnyTransaction?) async throws {
+    func delete0(within transaction: (any Transaction)?) async throws {
         try await Self.storage.delete(
             by: self.getIDAsKey(),
             within: transaction
@@ -108,19 +108,19 @@ public extension Entita2Entity {
     }
 
     /// This method is not intended to be used directly
-    func commit0(transaction: AnyTransaction?) async throws {
+    func commit0(transaction: (any Transaction)?) async throws {
         try await transaction?.commit()
     }
 
-    internal func commit0IfNecessary(commit: Bool, transaction: AnyTransaction?) async throws {
+    internal func commit0IfNecessary(commit: Bool, transaction: (any Transaction)?) async throws {
         if commit  {
             try await self.commit0(transaction: transaction)
         }
     }
 
     @usableFromInline
-    internal static func unwrapAnyTransactionOrBegin(_ anyTransaction: AnyTransaction?) async throws -> AnyTransaction {
-        if let transaction = anyTransaction {
+    internal static func unwrapTransactionOrBegin(_ transaction: (any Transaction)?) async throws -> any Transaction {
+        if let transaction {
             return transaction
         }
         return try await Self.storage.begin()
@@ -129,8 +129,8 @@ public extension Entita2Entity {
     // MARK: - Public CRUD methods
 
     /// Inserts current entity to DB within given transaction
-    func insert(within transaction: AnyTransaction? = nil, commit: Bool = true) async throws {
-        let transaction = try await Self.unwrapAnyTransactionOrBegin(transaction)
+    func insert(within transaction: (any Transaction)? = nil, commit: Bool = true) async throws {
+        let transaction = try await Self.unwrapTransactionOrBegin(transaction)
         
         try await self.beforeInsert0(within: transaction)
         try await self.beforeInsert(within: transaction)
@@ -141,13 +141,13 @@ public extension Entita2Entity {
     }
 
     /// Saves current entity to DB
-    func save(within transaction: AnyTransaction? = nil, commit: Bool = true) async throws {
+    func save(within transaction: (any Transaction)? = nil, commit: Bool = true) async throws {
         try await self.save(by: nil, within: transaction, commit: commit)
     }
 
     /// Saves current entity to DB within given transaction
-    func save(by ID: Identifier? = nil, within transaction: AnyTransaction? = nil, commit: Bool = true) async throws {
-        let transaction = try await Self.unwrapAnyTransactionOrBegin(transaction)
+    func save(by ID: Identifier? = nil, within transaction: (any Transaction)? = nil, commit: Bool = true) async throws {
+        let transaction = try await Self.unwrapTransactionOrBegin(transaction)
 
         try await self.beforeSave0(within: transaction)
         try await self.beforeSave(within: transaction)
@@ -158,8 +158,8 @@ public extension Entita2Entity {
     }
 
     /// Deletes current entity from DB within given transaction
-    func delete(within transaction: AnyTransaction? = nil, commit: Bool = true) async throws {
-        let transaction = try await Self.unwrapAnyTransactionOrBegin(transaction)
+    func delete(within transaction: (any Transaction)? = nil, commit: Bool = true) async throws {
+        let transaction = try await Self.unwrapTransactionOrBegin(transaction)
 
         try await self.beforeDelete0(within: transaction)
         try await self.beforeDelete(within: transaction)
