@@ -2,14 +2,14 @@ public typealias E2Storage = Entita2Storage
 
 public protocol Entita2Storage {
     /// Begins a transaction if `Storage` is transactional
-    func begin() async throws -> AnyTransaction
+    func begin() async throws -> any Transaction
 
     /// Tries to load bytes from storage for given key within a transaction
-    func load(by key: Bytes, within transaction: AnyTransaction?) async throws -> Bytes?
+    func load(by key: Bytes, within transaction: (any Transaction)?) async throws -> Bytes?
 
     /// Saves given bytes at given key within a transaction
-    func save(bytes: Bytes, by key: Bytes, within transaction: AnyTransaction?) async throws
+    func save(bytes: Bytes, by key: Bytes, within transaction: (any Transaction)?) async throws
 
     /// Deletes a given key (and value) from storage within a transaction
-    func delete(by key: Bytes, within transaction: AnyTransaction?) async throws
+    func delete(by key: Bytes, within transaction: (any Transaction)?) async throws
 }
